Moved app code into it's own folder to tidy things upa wee bit
