+++ .gitignore
@@ -245,3 +245,4 @@ ModelManifest.xml
 .fake/
 /Umb.LukeWarrenDev/App_Data/cache/2/d/c/b/8/9
 /Demo/App_Data/Logs
+/Demo/App_Data/TEMP
