+++ UmbracoGAEventTracking/UmbracoGAEventTracking.csproj
@@ -290,9 +290,9 @@
     <Content Include="Scripts\SendGaEvent.js" />
   </ItemGroup>
   <ItemGroup>
-    <Content Include="App_Code\HtmlExtensions.cs" />
-    <Content Include="App_Code\EventService.cs" />
-    <Content Include="App_Code\Event.cs" />
+    <Content Include="App_Code\GAEventTracking\HtmlExtensions.cs" />
+    <Content Include="App_Code\GAEventTracking\EventService.cs" />
+    <Content Include="App_Code\GAEventTracking\Event.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup />
@@ -330,9 +330,9 @@
   </Target>
   <Import Project="..\packages\AutoMapper.3.3.1\tools\AutoMapper.targets" Condition="Exists('..\packages\AutoMapper.3.3.1\tools\AutoMapper.targets')" />
   <PropertyGroup>
-    <PostBuildEvent>xcopy /E /Y "$(ProjectDir)Partials\*.cshtml" "$(SolutionDir)Demo\Views\Partials"
-xcopy /E /Y "$(ProjectDir)App_Code\*.cs" "$(SolutionDir)Demo\App_Code"
-xcopy /E /Y "$(ProjectDir)Scripts\*.js" "$(SolutionDir)Demo\Scripts"</PostBuildEvent>
+    <PostBuildEvent>xcopy /E /Y /S /D "$(ProjectDir)Partials\*.cshtml" "$(SolutionDir)Demo\Views\Partials"
+xcopy /E /Y /S /D "$(ProjectDir)App_Code\*.cs" "$(SolutionDir)Demo\App_Code"
+xcopy /E /Y /S /D "$(ProjectDir)Scripts\*.js" "$(SolutionDir)Demo\Scripts"</PostBuildEvent>
   </PropertyGroup>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
